requirements.txt
langgraph~=0.2.67
langchain_community~=0.3.25
langchain
python-dotenv~=1.1.0
ipykernel
langchain-openai~=0.3.22

langchain-core~=0.3.65
pydantic~=2.11.5
langgraph-cli
langgraph-runtime-inmem
langgraph-cli[inmem]
mcp
fastmcp
langchain-mcp-adapters